Address CI workflow review feedback
